puddle/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer 3.1.0
Django==4.1.7
django-paypal==2.0
idna==3.4
Pillow==9.4.0
pip==23.0.1
pipreqs==0.4.13
pytz==2023.3
requests==2.28.2
setuptools==58.1.0
sqlparse==0.4.3
tzdata==2022.7
Unidecode==1.3.6
urllib3==1.26.15
yarg==0.1.9